splitting visualization.py into 3 files
