+++ qiskit/tools/visualization/__init__.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+
+# Copyright 2018, IBM.
+#
+# This source code is licensed under the Apache License, Version 2.0 found in
+# the LICENSE.txt file in the root directory of this source tree.
+
+"""Main QISKit visualization methods."""
+
+from ._circuit_visualization import circuit_drawer, plot_circuit, latex_drawer, matplotlib_drawer
+from ._state_visualization import plot_state
+from ._counts_visualization import plot_histogram
