+++ qiskit/tools/visualization/_circuit_visualization.py
@@ -1,678 +1,60 @@
 # -*- coding: utf-8 -*-
 
-# Copyright 2017, IBM.
+# Copyright 2018, IBM.
 #
 # This source code is licensed under the Apache License, Version 2.0 found in
 # the LICENSE.txt file in the root directory of this source tree.
 
-# pylint: disable=invalid-name,anomalous-backslash-in-string
+# pylint: disable=invalid-name,anomalous-backslash-in-string,missing-docstring
 
 """
-Visualization functions for quantum states.
+Two quantum circuit drawers based on:
+    1. LaTeX
+    2. Matplotlib
 """
 
-import itertools
-import operator
-import re
-import os
-import subprocess
-import tempfile
-import logging
-import math
-from collections import Counter, OrderedDict
-from functools import reduce
+from collections import namedtuple, OrderedDict
+from fractions import Fraction
+from itertools import groupby, zip_longest
+from math import fmod, isclose, ceil
 from io import StringIO
 
-import numpy as np
-from scipy import linalg as la
+import re
+import os
+import operator
+import subprocess
+import logging
+import json
+import tempfile
+
+import matplotlib.patches as patches
 import matplotlib.pyplot as plt
-from matplotlib import cm
-from matplotlib.patches import FancyArrowPatch
-from mpl_toolkits.mplot3d import proj3d
+import numpy as np
+
+from qiskit import QuantumCircuit, QISKitError, load_qasm_file
+from qiskit.qasm import Qasm
+from qiskit.unroll import Unroller, JsonBackend
+
 try:
     from PIL import Image, ImageChops
 except ImportError:
     Image = None
     ImageChops = None
 
-from qiskit import qasm, unroll, QISKitError
-from qiskit.tools.qi.pauli import pauli_group, pauli_singles
 
 logger = logging.getLogger(__name__)
 
-###############################################################
-# Plotting histogram
-###############################################################
-
-
-def plot_histogram(data, number_to_keep=False):
-    """Plot a histogram of data.
-
-    data is a dictionary of  {'000': 5, '010': 113, ...}
-    number_to_keep is the number of terms to plot and rest is made into a
-    single bar called other values
-    """
-    if number_to_keep is not False:
-        data_temp = dict(Counter(data).most_common(number_to_keep))
-        data_temp["rest"] = sum(data.values()) - sum(data_temp.values())
-        data = data_temp
-
-    labels = sorted(data)
-    values = np.array([data[key] for key in labels], dtype=float)
-    pvalues = values / sum(values)
-    numelem = len(values)
-    ind = np.arange(numelem)  # the x locations for the groups
-    width = 0.35  # the width of the bars
-    _, ax = plt.subplots()
-    rects = ax.bar(ind, pvalues, width, color='seagreen')
-    # add some text for labels, title, and axes ticks
-    ax.set_ylabel('Probabilities', fontsize=12)
-    ax.set_xticks(ind)
-    ax.set_xticklabels(labels, fontsize=12, rotation=70)
-    ax.set_ylim([0., min([1.2, max([1.2 * val for val in pvalues])])])
-    # attach some text labels
-    for rect in rects:
-        height = rect.get_height()
-        ax.text(rect.get_x() + rect.get_width() / 2., 1.05 * height,
-                '%f' % float(height),
-                ha='center', va='bottom')
-    plt.show()
-
-
-###############################################################
-# Plotting states
-###############################################################
-
-class Arrow3D(FancyArrowPatch):
-    """Standard 3D arrow."""
-
-    def __init__(self, xs, ys, zs, *args, **kwargs):
-        """Create arrow."""
-        FancyArrowPatch.__init__(self, (0, 0), (0, 0), *args, **kwargs)
-        self._verts3d = xs, ys, zs
-
-    def draw(self, renderer):
-        """Draw the arrow."""
-        xs3d, ys3d, zs3d = self._verts3d
-        xs, ys, _ = proj3d.proj_transform(xs3d, ys3d, zs3d, renderer.M)
-        self.set_positions((xs[0], ys[0]), (xs[1], ys[1]))
-        FancyArrowPatch.draw(self, renderer)
-
-
-def plot_bloch_vector(bloch, title=""):
-    """Plot the Bloch sphere.
-
-    Plot a sphere, axes, the Bloch vector, and its projections onto each axis.
-
-    Args:
-        bloch (list[double]): array of three elements where [<x>, <y>,<z>]
-        title (str): a string that represents the plot title
-    """
-    # Set arrow lengths
-    arlen = 1.3
-
-    # Plot semi-transparent sphere
-    u = np.linspace(0, 2 * np.pi, 100)
-    v = np.linspace(0, np.pi, 100)
-    x = np.outer(np.cos(u), np.sin(v))
-    y = np.outer(np.sin(u), np.sin(v))
-    z = np.outer(np.ones(np.size(u)), np.cos(v))
-
-    fig = plt.figure(figsize=(6, 6))
-    ax = fig.add_subplot(111, projection='3d')
-    ax.set_aspect("equal")
-    ax.plot_surface(x, y, z, color=(.5, .5, .5), alpha=0.1)
-
-    # Plot arrows (axes, Bloch vector, its projections)
-    xa = Arrow3D([0, arlen], [0, 0], [0, 0], mutation_scale=20, lw=1,
-                 arrowstyle="-|>", color=(.5, .5, .5))
-    ya = Arrow3D([0, 0], [0, arlen], [0, 0], mutation_scale=20, lw=1,
-                 arrowstyle="-|>", color=(.5, .5, .5))
-    za = Arrow3D([0, 0], [0, 0], [0, arlen], mutation_scale=20, lw=1,
-                 arrowstyle="-|>", color=(.5, .5, .5))
-    a = Arrow3D([0, bloch[0]], [0, bloch[1]], [0, bloch[2]], mutation_scale=20,
-                lw=2, arrowstyle="simple", color="k")
-    bax = Arrow3D([0, bloch[0]], [0, 0], [0, 0], mutation_scale=20, lw=2,
-                  arrowstyle="-", color="r")
-    bay = Arrow3D([0, 0], [0, bloch[1]], [0, 0], mutation_scale=20, lw=2,
-                  arrowstyle="-", color="g")
-    baz = Arrow3D([0, 0], [0, 0], [0, bloch[2]], mutation_scale=20, lw=2,
-                  arrowstyle="-", color="b")
-    arrowlist = [xa, ya, za, a, bax, bay, baz]
-    for arr in arrowlist:
-        ax.add_artist(arr)
-
-    # Rotate the view
-    ax.view_init(30, 30)
-
-    # Annotate the axes, shifts are ad-hoc for this (30, 30) view
-    xp, yp, _ = proj3d.proj_transform(arlen, 0, 0, ax.get_proj())
-    plt.annotate("x", xy=(xp, yp), xytext=(-3, -8), textcoords='offset points',
-                 ha='right', va='bottom')
-    xp, yp, _ = proj3d.proj_transform(0, arlen, 0, ax.get_proj())
-    plt.annotate("y", xy=(xp, yp), xytext=(6, -5), textcoords='offset points',
-                 ha='right', va='bottom')
-    xp, yp, _ = proj3d.proj_transform(0, 0, arlen, ax.get_proj())
-    plt.annotate("z", xy=(xp, yp), xytext=(2, 0), textcoords='offset points',
-                 ha='right', va='bottom')
-
-    plt.title(title)
-    plt.show()
-
-
-def plot_state_city(rho, title=""):
-    """Plot the cityscape of quantum state.
-
-    Plot two 3d bargraphs (two dimenstional) of the mixed state rho
-
-    Args:
-        rho (np.array[[complex]]): array of dimensions 2**n x 2**nn complex
-                                   numbers
-        title (str): a string that represents the plot title
-    """
-    num = int(np.log2(len(rho)))
-
-    # get the real and imag parts of rho
-    datareal = np.real(rho)
-    dataimag = np.imag(rho)
-
-    # get the labels
-    column_names = [bin(i)[2:].zfill(num) for i in range(2**num)]
-    row_names = [bin(i)[2:].zfill(num) for i in range(2**num)]
-
-    lx = len(datareal[0])            # Work out matrix dimensions
-    ly = len(datareal[:, 0])
-    xpos = np.arange(0, lx, 1)    # Set up a mesh of positions
-    ypos = np.arange(0, ly, 1)
-    xpos, ypos = np.meshgrid(xpos+0.25, ypos+0.25)
-
-    xpos = xpos.flatten()
-    ypos = ypos.flatten()
-    zpos = np.zeros(lx*ly)
-
-    dx = 0.5 * np.ones_like(zpos)  # width of bars
-    dy = dx.copy()
-    dzr = datareal.flatten()
-    dzi = dataimag.flatten()
-
-    fig = plt.figure(figsize=(8, 8))
-    ax1 = fig.add_subplot(2, 1, 1, projection='3d')
-    ax1.bar3d(xpos, ypos, zpos, dx, dy, dzr, color="g", alpha=0.5)
-    ax2 = fig.add_subplot(2, 1, 2, projection='3d')
-    ax2.bar3d(xpos, ypos, zpos, dx, dy, dzi, color="g", alpha=0.5)
-
-    ax1.set_xticks(np.arange(0.5, lx+0.5, 1))
-    ax1.set_yticks(np.arange(0.5, ly+0.5, 1))
-    ax1.axes.set_zlim3d(-1.0, 1.0001)
-    ax1.set_zticks(np.arange(-1, 1, 0.5))
-    ax1.w_xaxis.set_ticklabels(row_names, fontsize=12, rotation=45)
-    ax1.w_yaxis.set_ticklabels(column_names, fontsize=12, rotation=-22.5)
-    # ax1.set_xlabel('basis state', fontsize=12)
-    # ax1.set_ylabel('basis state', fontsize=12)
-    ax1.set_zlabel("Real[rho]")
-
-    ax2.set_xticks(np.arange(0.5, lx+0.5, 1))
-    ax2.set_yticks(np.arange(0.5, ly+0.5, 1))
-    ax2.axes.set_zlim3d(-1.0, 1.0001)
-    ax2.set_zticks(np.arange(-1, 1, 0.5))
-    ax2.w_xaxis.set_ticklabels(row_names, fontsize=12, rotation=45)
-    ax2.w_yaxis.set_ticklabels(column_names, fontsize=12, rotation=-22.5)
-    # ax2.set_xlabel('basis state', fontsize=12)
-    # ax2.set_ylabel('basis state', fontsize=12)
-    ax2.set_zlabel("Imag[rho]")
-    plt.title(title)
-    plt.show()
-
-
-def plot_state_paulivec(rho, title=""):
-    """Plot the paulivec representation of a quantum state.
-
-    Plot a bargraph of the mixed state rho over the pauli matricies
-
-    Args:
-        rho (np.array[[complex]]): array of dimensions 2**n x 2**nn complex
-                                   numbers
-        title (str): a string that represents the plot title
-    """
-    num = int(np.log2(len(rho)))
-    labels = list(map(lambda x: x.to_label(), pauli_group(num)))
-    values = list(map(lambda x: np.real(np.trace(np.dot(x.to_matrix(), rho))),
-                      pauli_group(num)))
-    numelem = len(values)
-    ind = np.arange(numelem)  # the x locations for the groups
-    width = 0.5  # the width of the bars
-    _, ax = plt.subplots()
-    ax.grid(zorder=0)
-    ax.bar(ind, values, width, color='seagreen')
-
-    # add some text for labels, title, and axes ticks
-    ax.set_ylabel('Expectation value', fontsize=12)
-    ax.set_xticks(ind)
-    ax.set_yticks([-1, -0.5, 0, 0.5, 1])
-    ax.set_xticklabels(labels, fontsize=12, rotation=70)
-    ax.set_xlabel('Pauli', fontsize=12)
-    ax.set_ylim([-1, 1])
-    plt.title(title)
-    plt.show()
-
-
-def n_choose_k(n, k):
-    """Return the number of combinations for n choose k.
-
-    Args:
-        n (int): the total number of options .
-        k (int): The number of elements.
-
-    Returns:
-        int: returns the binomial coefficient
-    """
-    if n == 0:
-        return 0
-    return reduce(lambda x, y: x * y[0] / y[1],
-                  zip(range(n - k + 1, n + 1),
-                      range(1, k + 1)), 1)
-
-
-def lex_index(n, k, lst):
-    """Return  the lex index of a combination..
-
-    Args:
-        n (int): the total number of options .
-        k (int): The number of elements.
-        lst (list): list
-
-    Returns:
-        int: returns int index for lex order
-
-    """
-    assert len(lst) == k, "list should have length k"
-    comb = list(map(lambda x: n - 1 - x, lst))
-    dualm = sum([n_choose_k(comb[k - 1 - i], i + 1) for i in range(k)])
-    return int(dualm)
-
-
-def bit_string_index(s):
-    """Return the index of a string of 0s and 1s."""
-    n = len(s)
-    k = s.count("1")
-    assert s.count("0") == n - k, "s must be a string of 0 and 1"
-    ones = [pos for pos, char in enumerate(s) if char == "1"]
-    return lex_index(n, k, ones)
-
-
-def phase_to_color_wheel(complex_number):
-    """Map a phase of a complexnumber to a color in (r,g,b).
-
-    complex_number is phase is first mapped to angle in the range
-    [0, 2pi] and then to a color wheel with blue at zero phase.
-    """
-    angles = np.angle(complex_number)
-    angle_round = int(((angles + 2 * np.pi) % (2 * np.pi))/np.pi*6)
-    color_map = {
-        0: (0, 0, 1),  # blue,
-        1: (0.5, 0, 1),  # blue-violet
-        2: (1, 0, 1),  # violet
-        3: (1, 0, 0.5),  # red-violet,
-        4: (1, 0, 0),  # red
-        5: (1, 0.5, 0),  # red-oranage,
-        6: (1, 1, 0),  # orange
-        7: (0.5, 1, 0),  # orange-yellow
-        8: (0, 1, 0),  # yellow,
-        9: (0, 1, 0.5),  # yellow-green,
-        10: (0, 1, 1),  # green,
-        11: (0, 0.5, 1)  # green-blue,
-    }
-    return color_map[angle_round]
-
-
-def plot_state_qsphere(rho):
-    """Plot the qsphere representation of a quantum state."""
-    num = int(np.log2(len(rho)))
-    # get the eigenvectors and egivenvalues
-    we, stateall = la.eigh(rho)
-    for k in range(2**num):
-        # start with the max
-        probmix = we.max()
-        prob_location = we.argmax()
-        if probmix > 0.001:
-            print("The " + str(k) + "th eigenvalue = " + str(probmix))
-            # get the max eigenvalue
-            state = stateall[:, prob_location]
-            loc = np.absolute(state).argmax()
-            # get the element location closes to lowest bin representation.
-            for j in range(2**num):
-                test = np.absolute(np.absolute(state[j])
-                                   - np.absolute(state[loc]))
-                if test < 0.001:
-                    loc = j
-                    break
-            # remove the global phase
-            angles = (np.angle(state[loc]) + 2 * np.pi) % (2 * np.pi)
-            angleset = np.exp(-1j*angles)
-            # print(state)
-            # print(angles)
-            state = angleset*state
-            # print(state)
-            state.flatten()
-            # start the plotting
-            fig = plt.figure(figsize=(10, 10))
-            ax = fig.add_subplot(111, projection='3d')
-            ax.axes.set_xlim3d(-1.0, 1.0)
-            ax.axes.set_ylim3d(-1.0, 1.0)
-            ax.axes.set_zlim3d(-1.0, 1.0)
-            ax.set_aspect("equal")
-            ax.axes.grid(False)
-            # Plot semi-transparent sphere
-            u = np.linspace(0, 2 * np.pi, 25)
-            v = np.linspace(0, np.pi, 25)
-            x = np.outer(np.cos(u), np.sin(v))
-            y = np.outer(np.sin(u), np.sin(v))
-            z = np.outer(np.ones(np.size(u)), np.cos(v))
-            ax.plot_surface(x, y, z, rstride=1, cstride=1, color='k',
-                            alpha=0.05, linewidth=0)
-            # wireframe
-            # Get rid of the panes
-            ax.w_xaxis.set_pane_color((1.0, 1.0, 1.0, 0.0))
-            ax.w_yaxis.set_pane_color((1.0, 1.0, 1.0, 0.0))
-            ax.w_zaxis.set_pane_color((1.0, 1.0, 1.0, 0.0))
-
-            # Get rid of the spines
-            ax.w_xaxis.line.set_color((1.0, 1.0, 1.0, 0.0))
-            ax.w_yaxis.line.set_color((1.0, 1.0, 1.0, 0.0))
-            ax.w_zaxis.line.set_color((1.0, 1.0, 1.0, 0.0))
-            # Get rid of the ticks
-            ax.set_xticks([])
-            ax.set_yticks([])
-            ax.set_zticks([])
-
-            d = num
-            for i in range(2**num):
-                # get x,y,z points
-                element = bin(i)[2:].zfill(num)
-                weight = element.count("1")
-                zvalue = -2 * weight / d + 1
-                number_of_divisions = n_choose_k(d, weight)
-                weight_order = bit_string_index(element)
-                # if weight_order >= number_of_divisions / 2:
-                #    com_key = compliment(element)
-                #    weight_order_temp = bit_string_index(com_key)
-                #    weight_order = np.floor(
-                #        number_of_divisions / 2) + weight_order_temp + 1
-                angle = weight_order * 2 * np.pi / number_of_divisions
-                xvalue = np.sqrt(1 - zvalue**2) * np.cos(angle)
-                yvalue = np.sqrt(1 - zvalue**2) * np.sin(angle)
-                ax.plot([xvalue], [yvalue], [zvalue],
-                        markerfacecolor=(.5, .5, .5),
-                        markeredgecolor=(.5, .5, .5),
-                        marker='o', markersize=10, alpha=1)
-                # get prob and angle - prob will be shade and angle color
-                prob = np.real(np.dot(state[i], state[i].conj()))
-                colorstate = phase_to_color_wheel(state[i])
-                a = Arrow3D([0, xvalue], [0, yvalue], [0, zvalue],
-                            mutation_scale=20, alpha=prob, arrowstyle="-",
-                            color=colorstate, lw=10)
-                ax.add_artist(a)
-            # add weight lines
-            for weight in range(d + 1):
-                theta = np.linspace(-2 * np.pi, 2 * np.pi, 100)
-                z = -2 * weight / d + 1
-                r = np.sqrt(1 - z**2)
-                x = r * np.cos(theta)
-                y = r * np.sin(theta)
-                ax.plot(x, y, z, color=(.5, .5, .5))
-            # add center point
-            ax.plot([0], [0], [0], markerfacecolor=(.5, .5, .5),
-                    markeredgecolor=(.5, .5, .5), marker='o', markersize=10,
-                    alpha=1)
-            plt.show()
-            we[prob_location] = 0
-        else:
-            break
-
-
-def plot_state(rho, method='city'):
-    """Plot the quantum state."""
-    num = int(np.log2(len(rho)))
-    # Need updating to check its a matrix
-    if method == 'city':
-        plot_state_city(rho)
-    elif method == "paulivec":
-        plot_state_paulivec(rho)
-    elif method == "qsphere":
-        plot_state_qsphere(rho)
-    elif method == "bloch":
-        for i in range(num):
-            bloch_state = list(
-                map(lambda x: np.real(np.trace(np.dot(x.to_matrix(), rho))),
-                    pauli_singles(i, num)))
-            plot_bloch_vector(bloch_state, "qubit " + str(i))
-    elif method == "wigner":
-        plot_wigner_function(rho)
-
-
-###############################################################
-# Plotting Wigner functions
-###############################################################
-
-def plot_wigner_function(state, res=100):
-    """Plot the equal angle slice spin Wigner function of an arbitrary
-    quantum state.
-
-    Args:
-        state (np.matrix[[complex]]):
-            - Matrix of 2**n x 2**n complex numbers
-            - State Vector of 2**n x 1 complex numbers
-        res (int) : number of theta and phi values in meshgrid
-            on sphere (creates a res x res grid of points)
-
-    References:
-        [1] T. Tilma, M. J. Everitt, J. H. Samson, W. J. Munro,
-        and K. Nemoto, Phys. Rev. Lett. 117, 180401 (2016).
-        [2] R. P. Rundle, P. W. Mills, T. Tilma, J. H. Samson, and
-        M. J. Everitt, Phys. Rev. A 96, 022117 (2017).
-    """
-    state = np.array(state)
-    if state.ndim == 1:
-        state = np.outer(state,
-                         state)  # turns state vector to a density matrix
-    state = np.matrix(state)
-    num = int(np.log2(len(state)))  # number of qubits
-    phi_vals = np.linspace(0, np.pi, num=res,
-                           dtype=np.complex_)
-    theta_vals = np.linspace(0, 0.5*np.pi, num=res,
-                             dtype=np.complex_)  # phi and theta values for WF
-    w = np.empty([res, res])
-    harr = np.sqrt(3)
-    delta_su2 = np.zeros((2, 2), dtype=np.complex_)
-
-    # create the spin Wigner function
-    for theta in range(res):
-        costheta = harr*np.cos(2*theta_vals[theta])
-        sintheta = harr*np.sin(2*theta_vals[theta])
-
-        for phi in range(res):
-            delta_su2[0, 0] = 0.5*(1+costheta)
-            delta_su2[0, 1] = -0.5*(np.exp(2j*phi_vals[phi])*sintheta)
-            delta_su2[1, 0] = -0.5*(np.exp(-2j*phi_vals[phi])*sintheta)
-            delta_su2[1, 1] = 0.5*(1-costheta)
-            kernel = 1
-            for _ in range(num):
-                kernel = np.kron(kernel,
-                                 delta_su2)  # creates phase point kernel
-
-            w[phi, theta] = np.real(np.trace(state*kernel))  # Wigner function
-
-    # Plot a sphere (x,y,z) with Wigner function facecolor data stored in Wc
-    fig = plt.figure(figsize=(11, 9))
-    ax = fig.gca(projection='3d')
-    w_max = np.amax(w)
-    # Color data for plotting
-    w_c = cm.seismic_r((w+w_max)/(2*w_max))  # color data for sphere
-    w_c2 = cm.seismic_r((w[0:res, int(res/2):res]+w_max)/(2*w_max))  # bottom
-    w_c3 = cm.seismic_r((w[int(res/4):int(3*res/4), 0:res]+w_max) /
-                        (2*w_max))  # side
-    w_c4 = cm.seismic_r((w[int(res/2):res, 0:res]+w_max)/(2*w_max))  # back
-
-    u = np.linspace(0, 2 * np.pi, res)
-    v = np.linspace(0, np.pi, res)
-    x = np.outer(np.cos(u), np.sin(v))
-    y = np.outer(np.sin(u), np.sin(v))
-    z = np.outer(np.ones(np.size(u)), np.cos(v))  # creates a sphere mesh
-
-    ax.plot_surface(x, y, z, facecolors=w_c,
-                    vmin=-w_max, vmax=w_max,
-                    rcount=res, ccount=res,
-                    linewidth=0, zorder=0.5,
-                    antialiased=False)  # plots Wigner Bloch sphere
-
-    ax.plot_surface(x[0:res, int(res/2):res],
-                    y[0:res, int(res/2):res],
-                    -1.5*np.ones((res, int(res/2))),
-                    facecolors=w_c2,
-                    vmin=-w_max, vmax=w_max,
-                    rcount=res/2, ccount=res/2,
-                    linewidth=0, zorder=0.5,
-                    antialiased=False)  # plots bottom reflection
-
-    ax.plot_surface(-1.5*np.ones((int(res/2), res)),
-                    y[int(res/4):int(3*res/4), 0:res],
-                    z[int(res/4):int(3*res/4), 0:res],
-                    facecolors=w_c3,
-                    vmin=-w_max, vmax=w_max,
-                    rcount=res/2, ccount=res/2,
-                    linewidth=0, zorder=0.5,
-                    antialiased=False)  # plots side reflection
-
-    ax.plot_surface(x[int(res/2):res, 0:res],
-                    1.5*np.ones((int(res/2), res)),
-                    z[int(res/2):res, 0:res],
-                    facecolors=w_c4,
-                    vmin=-w_max, vmax=w_max,
-                    rcount=res/2, ccount=res/2,
-                    linewidth=0, zorder=0.5,
-                    antialiased=False)  # plots back reflection
-
-    ax.w_xaxis.set_pane_color((0.4, 0.4, 0.4, 1.0))
-    ax.w_yaxis.set_pane_color((0.4, 0.4, 0.4, 1.0))
-    ax.w_zaxis.set_pane_color((0.4, 0.4, 0.4, 1.0))
-    ax.set_xticks([], [])
-    ax.set_yticks([], [])
-    ax.set_zticks([], [])
-    ax.grid(False)
-    ax.xaxis.pane.set_edgecolor('black')
-    ax.yaxis.pane.set_edgecolor('black')
-    ax.zaxis.pane.set_edgecolor('black')
-    ax.set_xlim(-1.5, 1.5)
-    ax.set_ylim(-1.5, 1.5)
-    ax.set_zlim(-1.5, 1.5)
-    m = cm.ScalarMappable(cmap=cm.seismic_r)
-    m.set_array([-w_max, w_max])
-    plt.colorbar(m, shrink=0.5, aspect=10)
-
-    plt.show()
-
-
-def plot_wigner_curve(wigner_data, xaxis=None):
-    """Plots a curve for points in phase space of the spin Wigner function.
-
-    Args:
-        wigner_data(np.array): an array of points to plot as a 2d curve
-        xaxis (np.array):  the range of the x axis
-    """
-    if not xaxis:
-        xaxis = np.linspace(0, len(wigner_data)-1, num=len(wigner_data))
-
-    plt.plot(xaxis, wigner_data)
-    plt.show()
-
-
-def plot_wigner_plaquette(wigner_data, max_wigner='local'):
-    """Plots plaquette of wigner function data, the plaquette will
-    consist of cicles each colored to match the value of the Wigner
-    function at the given point in phase space.
-
-    Args:
-        wigner_data (matrix): array of Wigner function data where the
-                            rows are plotted along the x axis and the
-                            columns are plotted along the y axis
-        max_wigner (str or float):
-            - 'local' puts the maximum value to maximum of the points
-            - 'unit' sets maximum to 1
-            - float for a custom maximum.
-    """
-    wigner_data = np.matrix(wigner_data)
-    dim = wigner_data.shape
-
-    if max_wigner == 'local':
-        w_max = np.amax(wigner_data)
-    elif max_wigner == 'unit':
-        w_max = 1
-    else:
-        w_max = max_wigner  # For a float input
-    w_max = float(w_max)
-
-    cmap = plt.cm.get_cmap('seismic_r')
-
-    xax = dim[1]-0.5
-    yax = dim[0]-0.5
-    norm = np.amax(dim)
-
-    fig = plt.figure(figsize=((xax+0.5)*6/norm, (yax+0.5)*6/norm))
-    ax = fig.gca()
-
-    for x in range(int(dim[1])):
-        for y in range(int(dim[0])):
-            circle = plt.Circle(
-                (x, y), 0.49, color=cmap((wigner_data[y, x]+w_max)/(2*w_max)))
-            ax.add_artist(circle)
-
-    ax.set_xlim(-1, xax+0.5)
-    ax.set_ylim(-1, yax+0.5)
-    ax.set_xticks([], [])
-    ax.set_yticks([], [])
-    m = cm.ScalarMappable(cmap=cm.seismic_r)
-    m.set_array([-w_max, w_max])
-    plt.colorbar(m, shrink=0.5, aspect=10)
-    plt.show()
-
-
-def plot_wigner_data(wigner_data, phis=None, method=None):
-    """Plots Wigner results in appropriate format.
-
-    Args:
-        wigner_data (numpy.array): Output returned from the wigner_data
-            function
-        phis (numpy.array): Values of phi
-        method (str or None): how the data is to be plotted, methods are:
-            point: a single point in phase space
-            curve: a two dimensional curve
-            plaquette: points plotted as circles
-    """
-    if not method:
-        wig_dim = len(np.shape(wigner_data))
-        if wig_dim == 1:
-            if np.shape(wigner_data) == 1:
-                method = 'point'
-            else:
-                method = 'curve'
-        elif wig_dim == 2:
-            method = 'plaquette'
-
-    if method == 'curve':
-        plot_wigner_curve(wigner_data, xaxis=phis)
-    elif method == 'plaquette':
-        plot_wigner_plaquette(wigner_data)
-    elif method == 'state':
-        plot_wigner_function(wigner_data)
-    elif method == 'point':
-        plot_wigner_plaquette(wigner_data)
-        print('point in phase space is '+str(wigner_data))
-    else:
-        print("No method given")
-
-###############################################################
-# Plotting circuit
-###############################################################
+# -----------------------------------------------------------------------------
+# definitions
+# -----------------------------------------------------------------------------
+WID = 0.65
+HIG = 0.65
+DEFAULT_SCALE = 3
+PORDER_GATE = 5
+PORDER_LINE = 2
+PORDER_GRAY = 3
+PORDER_TEXT = 6
+PORDER_SUBP = 4
 
 
 def plot_circuit(circuit,
@@ -777,11 +159,11 @@ def latex_drawer(circuit, filename=None,
     Returns:
         str: Latex string appropriate for writing to file.
     """
-    ast = qasm.Qasm(data=circuit.qasm()).parse()
+    ast = Qasm(data=circuit.qasm()).parse()
     if basis:
         # Split basis only if it is not the empty string.
         basis = basis.split(',')
-    u = unroll.Unroller(ast, unroll.JsonBackend(basis))
+    u = Unroller(ast, JsonBackend(basis))
     u.execute()
     json_circuit = u.backend.circuit
     qcimg = QCircuitImage(json_circuit, scale)
@@ -1177,7 +559,7 @@ class QCircuitImage(object):
         # the qubit/cbit labels plus initial states is 2 more
         # the wires poking out at the ends is 2 more
         sum_column_widths = sum(1 + v / 3 for v in max_column_width.values())
-        return columns+1, math.ceil(sum_column_widths)+4
+        return columns+1, ceil(sum_column_widths)+4
 
     def _get_beamer_page(self):
         """Get height, width & scale attributes for the beamer page.
@@ -1657,7 +1039,7 @@ def _get_register_specs(bit_labels):
     Yields:
         tuple: iterator of register_name:size pairs.
     """
-    it = itertools.groupby(bit_labels, operator.itemgetter(0))
+    it = groupby(bit_labels, operator.itemgetter(0))
     for register_name, sub_it in it:
         yield register_name, max(ind[1] for ind in sub_it) + 1
 
@@ -1674,3 +1056,742 @@ def _truncate_float(matchobj, format_str='0.2g'):
     if matchobj.group(0):
         return format(float(matchobj.group(0)), format_str)
     return ''
+
+
+def matplotlib_drawer(circuit,
+                      basis='id,u0,u1,u2,u3,x,y,z,h,s,sdg,t,tdg,rx,ry,rz,'
+                            'cx,cy,cz,ch,crz,cu1,cu3,swap,ccx,cswap',
+                      scale=1.0, style=None, filename=None):
+    """Draw a quantum circuit based on matplotlib.
+    If `%matplotlib inline` is invoked in a Jupyter notebook, it visualizes a circuit inline.
+    We recommend `%config InlineBackend.figure_format = 'svg'` for the inline visualization.
+
+    Args:
+        circuit (QuantumCircuit): a quantum circuit
+        basis (str): comma separated list of gates
+        scale (float): scaling factor
+        style (dict or str): dictionary of style or file name of style file
+        filename (str): output filename of circuit drawing if filename is not None
+    """
+    if ',' not in basis:
+        logger.warning('Warning: basis is not comma separated: "%s". '
+                       'Perhaps you set `filename` to `basis`.', basis)
+    qcd = MatplotlibDrawer(basis=basis, scale=scale, style=style)
+    qcd.parse_circuit(circuit)
+    qcd.draw(filename)
+
+
+Register = namedtuple('Register', 'name index')
+
+
+class QCStyle:
+    def __init__(self):
+        self.tc = '#000000'
+        self.sc = '#000000'
+        self.lc = '#000000'
+        self.cc = '#778899'
+        self.gc = '#ffffff'
+        self.gt = '#000000'
+        self.bc = '#bdbdbd'
+        self.bg = '#ffffff'
+        self.fs = 10
+        self.sfs = 6
+        self.disptex = {
+            'id': 'id',
+            'u0': 'U_0',
+            'u1': 'U_1',
+            'u2': 'U_2',
+            'u3': 'U_3',
+            'x': 'X',
+            'y': 'Y',
+            'z': 'Z',
+            'h': 'H',
+            's': 'S',
+            'sdg': 'S^\\dagger',
+            't': 'T',
+            'tdg': 'T^\\dagger',
+            'rx': 'R_x',
+            'ry': 'R_y',
+            'rz': 'R_z',
+            'reset': '\\left|0\\right\\rangle'
+        }
+        self.dispcol = {
+            'id': '#ffffff',
+            'u0': '#ffffff',
+            'u1': '#ffffff',
+            'u2': '#ffffff',
+            'u3': '#ffffff',
+            'x': '#ffffff',
+            'y': '#ffffff',
+            'z': '#ffffff',
+            'h': '#ffffff',
+            's': '#ffffff',
+            'sdg': '#ffffff',
+            't': '#ffffff',
+            'tdg': '#ffffff',
+            'rx': '#ffffff',
+            'ry': '#ffffff',
+            'rz': '#ffffff',
+            'reset': '#ffffff',
+            'target': '#ffffff',
+            'meas': '#ffffff'
+        }
+        self.latexmode = True
+        self.pimode = False
+        self.fold = 20
+        self.bundle = False
+        self.barrier = False
+        self.index = False
+        self.compress = False
+        self.figwidth = -1
+        self.dpi = 150
+
+    def set_style(self, dic):
+        self.tc = dic.get('textcolor', self.tc)
+        self.sc = dic.get('subtextcolor', self.sc)
+        self.lc = dic.get('linecolor', self.lc)
+        self.cc = dic.get('creglinecolor', self.cc)
+        self.gt = dic.get('gatetextcolor', self.tc)
+        self.gc = dic.get('gatefacecolor', self.gc)
+        self.bc = dic.get('barrierfacecolor', self.bc)
+        self.bg = dic.get('backgroundcolor', self.bg)
+        self.fs = dic.get('fontsize', self.fs)
+        self.sfs = dic.get('subfontsize', self.sfs)
+        self.disptex = dic.get('displaytext', self.disptex)
+        for key in self.dispcol.keys():
+            self.dispcol[key] = self.gc
+        self.dispcol = dic.get('displaycolor', self.dispcol)
+        self.latexmode = dic.get('latexdrawerstyle', self.latexmode)
+        self.pimode = dic.get('usepiformat', self.pimode)
+        self.fold = dic.get('fold', self.fold)
+        if self.fold < 2:
+            self.fold = -1
+        self.bundle = dic.get('cregbundle', self.bundle)
+        self.barrier = dic.get('plotbarrier', self.barrier)
+        self.index = dic.get('showindex', self.index)
+        self.compress = dic.get('compress', self.compress)
+        self.figwidth = dic.get('figwidth', self.figwidth)
+        self.dpi = dic.get('dpi', self.dpi)
+
+
+class Anchor:
+    def __init__(self, reg_num, yind, fold):
+        self.__yind = yind
+        self.__fold = fold
+        self.__reg_num = reg_num
+        self.__gate_placed = []
+
+    def plot_coord(self, index, gate_width):
+        h_pos = index % self.__fold + 1
+        # check folding
+        if self.__fold > 0:
+            if h_pos + (gate_width - 1) > self.__fold:
+                index += self.__fold - (h_pos - 1)
+            x_pos = index % self.__fold + 1 + 0.5 * (gate_width - 1)
+            y_pos = self.__yind - (index // self.__fold) * (self.__reg_num + 1)
+        else:
+            x_pos = index + 1 + 0.5 * (gate_width - 1)
+            y_pos = self.__yind
+
+        return x_pos, y_pos
+
+    def is_locatable(self, index, gate_width):
+        hold = [index + i for i in range(gate_width)]
+        for p in hold:
+            if p in self.__gate_placed:
+                return False
+        return True
+
+    def set_index(self, index, gate_width):
+        h_pos = index % self.__fold + 1
+        if h_pos + (gate_width - 1) > self.__fold:
+            _index = index + self.__fold - (h_pos - 1)
+        else:
+            _index = index
+        for ii in range(gate_width):
+            if _index + ii not in self.__gate_placed:
+                self.__gate_placed.append(_index + ii)
+        self.__gate_placed.sort()
+
+    def get_index(self):
+        if self.__gate_placed:
+            return self.__gate_placed[-1] + 1
+        return 0
+
+
+class MatplotlibDrawer:
+    def __init__(self,
+                 basis='id,u0,u1,u2,u3,x,y,z,h,s,sdg,t,tdg,rx,ry,rz,'
+                       'cx,cy,cz,ch,crz,cu1,cu3,swap,ccx,cswap',
+                 scale=1.0, style=None):
+
+        self._ast = None
+        self._basis = basis.split(',')
+        self._scale = DEFAULT_SCALE * scale
+        self._creg = []
+        self._qreg = []
+        self._ops = []
+        self._qreg_dict = {}
+        self._creg_dict = {}
+        self._cond = {
+            'n_lines': 0,
+            'xmax': 0,
+            'ymax': 0,
+        }
+
+        self._style = QCStyle()
+        if style:
+            if isinstance(style, dict):
+                self._style.set_style(style)
+            elif isinstance(style, str):
+                with open(style, 'r') as infile:
+                    dic = json.load(infile)
+                self._style.set_style(dic)
+
+        self.figure = plt.figure()
+        self.figure.patch.set_facecolor(color=self._style.bg)
+        self.ax = self.figure.add_subplot(111)
+        self.ax.axis('off')
+        self.ax.set_aspect('equal', 'datalim')
+
+    def load_qasm_file(self, filename):
+        circuit = load_qasm_file(filename, name='draw', basis_gates=','.join(self._basis))
+        self.parse_circuit(circuit)
+
+    def parse_circuit(self, circuit: QuantumCircuit):
+        ast = Qasm(data=circuit.qasm()).parse()
+        u = Unroller(ast, JsonBackend(self._basis))
+        u.execute()
+        self._ast = u.backend.circuit
+        self._registers()
+        self._ops = self._ast['operations']
+
+    def _registers(self):
+        # NOTE: formats of clbit and qubit are different!
+        header = self._ast['header']
+        self._creg = []
+        for e in header['clbit_labels']:
+            for i in range(e[1]):
+                self._creg.append(Register(name=e[0], index=i))
+        assert len(self._creg) == header['number_of_clbits']
+        self._qreg = []
+        for e in header['qubit_labels']:
+            self._qreg.append(Register(name=e[0], index=e[1]))
+        assert len(self._qreg) == header['number_of_qubits']
+
+    @property
+    def ast(self):
+        return self._ast
+
+    def _gate(self, xy, fc=None, wide=False, text=None, subtext=None):
+        xpos, ypos = xy
+
+        if wide:
+            wid = WID * 2.8
+        else:
+            wid = WID
+        if fc:
+            _fc = fc
+        elif text:
+            _fc = self._style.dispcol[text]
+        else:
+            _fc = self._style.gc
+
+        box = patches.Rectangle(xy=(xpos - 0.5 * wid, ypos - 0.5 * HIG), width=wid, height=HIG,
+                                fc=_fc, ec=self._style.lc, linewidth=1.5, zorder=PORDER_GATE)
+        self.ax.add_patch(box)
+
+        if text:
+            disp_text = "${}$".format(self._style.disptex[text])
+            if subtext:
+                self.ax.text(xpos, ypos + 0.15 * HIG, disp_text, ha='center', va='center',
+                             fontsize=self._style.fs,
+                             color=self._style.gt,
+                             zorder=PORDER_TEXT)
+                self.ax.text(xpos, ypos - 0.3 * HIG, subtext, ha='center', va='center',
+                             fontsize=self._style.sfs,
+                             color=self._style.sc,
+                             zorder=PORDER_TEXT)
+            else:
+                self.ax.text(xpos, ypos, disp_text, ha='center', va='center',
+                             fontsize=self._style.fs,
+                             color=self._style.gt,
+                             zorder=PORDER_TEXT)
+
+    def _subtext(self, xy, text):
+        xpos, ypos = xy
+
+        self.ax.text(xpos, ypos - 0.3 * HIG, text, ha='center', va='top',
+                     fontsize=self._style.sfs,
+                     color=self._style.tc,
+                     zorder=PORDER_TEXT)
+
+    def _line(self, xy0, xy1):
+        x0, y0 = xy0
+        x1, y1 = xy1
+        self.ax.plot([x0, x1], [y0, y1],
+                     color=self._style.lc,
+                     linewidth=1.0,
+                     zorder=PORDER_LINE)
+
+    def _measure(self, qxy, cxy, cid):
+        qx, qy = qxy
+        cx, cy = cxy
+
+        self._gate(qxy, fc=self._style.dispcol['meas'])
+        # add measure symbol
+        arc = patches.Arc(xy=(qx, qy - 0.15 * HIG), width=WID * 0.7, height=HIG * 0.7,
+                          theta1=0, theta2=180, fill=False,
+                          ec=self._style.lc, linewidth=1.5, zorder=PORDER_GATE)
+        self.ax.add_patch(arc)
+        self.ax.plot([qx, qx + 0.35 * WID], [qy - 0.15 * HIG, qy + 0.20 * HIG],
+                     color=self._style.lc, linewidth=1.5, zorder=PORDER_GATE)
+        # arrow
+        self.ax.arrow(x=qx, y=qy, dx=0, dy=cy - qy, width=0.01, head_width=0.2, head_length=0.2,
+                      length_includes_head=True, color=self._style.cc, zorder=PORDER_LINE)
+        # target
+        if self._style.bundle:
+            self.ax.text(cx + .25, cy + .1, str(cid), ha='left', va='bottom',
+                         fontsize=0.8 * self._style.fs,
+                         color=self._style.tc,
+                         zorder=PORDER_TEXT)
+
+    def _conds(self, xy, istrue=False):
+        xpos, ypos = xy
+
+        if istrue:
+            _fc = self._style.lc
+        else:
+            _fc = self._style.gc
+
+        box = patches.Circle(xy=(xpos, ypos), radius=WID * 0.15,
+                             fc=_fc, ec=self._style.lc,
+                             linewidth=1.5, zorder=PORDER_GATE)
+        self.ax.add_patch(box)
+
+    def _ctrl_qubit(self, xy):
+        xpos, ypos = xy
+
+        box = patches.Circle(xy=(xpos, ypos), radius=WID * 0.15,
+                             fc=self._style.lc, ec=self._style.lc,
+                             linewidth=1.5, zorder=PORDER_GATE)
+        self.ax.add_patch(box)
+
+    def _tgt_qubit(self, xy):
+        xpos, ypos = xy
+
+        box = patches.Circle(xy=(xpos, ypos), radius=HIG * 0.35,
+                             fc=self._style.dispcol['target'], ec=self._style.lc,
+                             linewidth=1.5, zorder=PORDER_GATE)
+        self.ax.add_patch(box)
+        # add '+' symbol
+        self.ax.plot([xpos, xpos], [ypos - 0.35 * HIG, ypos + 0.35 * HIG],
+                     color=self._style.lc, linewidth=1.0, zorder=PORDER_GATE)
+        self.ax.plot([xpos - 0.35 * HIG, xpos + 0.35 * HIG], [ypos, ypos],
+                     color=self._style.lc, linewidth=1.0, zorder=PORDER_GATE)
+
+    def _swap(self, xy):
+        xpos, ypos = xy
+
+        self.ax.plot([xpos - 0.20 * WID, xpos + 0.20 * WID], [ypos - 0.20 * WID, ypos + 0.20 * WID],
+                     color=self._style.lc, linewidth=1.5, zorder=PORDER_LINE)
+        self.ax.plot([xpos - 0.20 * WID, xpos + 0.20 * WID], [ypos + 0.20 * WID, ypos - 0.20 * WID],
+                     color=self._style.lc, linewidth=1.5, zorder=PORDER_LINE)
+
+    def _barrier(self, config, anc):
+        xys = config['coord']
+        group = config['group']
+        y_reg = []
+        for qreg in self._qreg_dict.values():
+            if qreg['group'] in group:
+                y_reg.append(qreg['y'])
+        x0 = xys[0][0]
+
+        box_y0 = min(y_reg) - int(anc / self._style.fold) * (self._cond['n_lines'] + 1) - 0.5
+        box_y1 = max(y_reg) - int(anc / self._style.fold) * (self._cond['n_lines'] + 1) + 0.5
+        box = patches.Rectangle(xy=(x0 - 0.3 * WID, box_y0),
+                                width=0.6 * WID, height=box_y1 - box_y0,
+                                fc=self._style.bc, ec=None, alpha=0.6,
+                                linewidth=1.5, zorder=PORDER_GRAY)
+        self.ax.add_patch(box)
+        for xy in xys:
+            xpos, ypos = xy
+            self.ax.plot([xpos, xpos], [ypos + 0.5, ypos - 0.5],
+                         linewidth=1, linestyle="dashed",
+                         color=self._style.lc,
+                         zorder=PORDER_TEXT)
+
+    def _linefeed_mark(self, xy):
+        xpos, ypos = xy
+
+        self.ax.plot([xpos - .1, xpos - .1],
+                     [ypos, ypos - self._cond['n_lines'] + 1],
+                     color=self._style.lc, zorder=PORDER_LINE)
+        self.ax.plot([xpos + .1, xpos + .1],
+                     [ypos, ypos - self._cond['n_lines'] + 1],
+                     color=self._style.lc, zorder=PORDER_LINE)
+
+    def draw(self, filename=None, verbose=False):
+        self._draw_regs()
+        self._draw_ops(verbose)
+        self.ax.set_xlim(-1.5, self._cond['xmax'] + 1.5)
+        self.ax.set_ylim(self._cond['ymax'] - 1.5, 1.5)
+        # update figure size
+        fig_w = abs(self._cond['xmax']) + 2
+        fig_h = abs(self._cond['ymax']) + 2
+        if self._style.figwidth < 0.0:
+            self._style.figwidth = fig_w * self._scale * self._style.fs / 72 / WID
+        self.figure.set_size_inches(self._style.figwidth, self._style.figwidth * fig_h / fig_w)
+        if filename:
+            self.figure.savefig(filename, dpi=self._style.dpi, bbox_inches='tight')
+        else:
+            plt.show()
+
+    def _draw_regs(self):
+        # quantum register
+        for ii, reg in enumerate(self._qreg):
+            if len(self._qreg) > 1:
+                label = '${}_{{{}}}$'.format(reg.name, reg.index)
+            else:
+                label = '${}$'.format(reg.name)
+            pos = -ii
+            self._qreg_dict[ii] = {'y': pos, 'label': label, 'index': reg.index, 'group': reg.name}
+            self._cond['n_lines'] += 1
+        # classical register
+        if self._creg:
+            n_creg = self._creg.copy()
+            n_creg.pop(0)
+            idx = 0
+            y_off = -len(self._qreg)
+            for ii, (reg, nreg) in enumerate(zip_longest(self._creg, n_creg)):
+                pos = y_off - idx
+                if self._style.bundle:
+                    label = '${}$'.format(reg.name)
+                    self._creg_dict[ii] = {'y': pos, 'label': label, 'index': reg.index,
+                                           'group': reg.name}
+                    if not (not nreg or reg.name != nreg.name):
+                        continue
+                else:
+                    label = '${}_{{{}}}$'.format(reg.name, reg.index)
+                    self._creg_dict[ii] = {'y': pos, 'label': label, 'index': reg.index,
+                                           'group': reg.name}
+                self._cond['n_lines'] += 1
+                idx += 1
+
+    def _draw_regs_sub(self, n_fold, feedline_l=False, feedline_r=False):
+        # quantum register
+        for qreg in self._qreg_dict.values():
+            if n_fold == 0:
+                label = qreg['label'] + ' : $\\left|0\\right\\rangle$'
+            else:
+                label = qreg['label']
+            y = qreg['y'] - n_fold * (self._cond['n_lines'] + 1)
+            self.ax.text(-0.5, y, label, ha='right', va='center',
+                         fontsize=self._style.fs,
+                         color=self._style.tc,
+                         zorder=PORDER_TEXT)
+            self.ax.plot([0, self._cond['xmax']], [y, y], color=self._style.lc, zorder=PORDER_LINE)
+        # classical register
+        this_creg_dict = {}
+        for creg in self._creg_dict.values():
+            if n_fold == 0:
+                label = creg['label'] + ' :  0 '
+            else:
+                label = creg['label']
+            y = creg['y'] - n_fold * (self._cond['n_lines'] + 1)
+            if y not in this_creg_dict.keys():
+                this_creg_dict[y] = {'val': 1, 'label': label}
+            else:
+                this_creg_dict[y]['val'] += 1
+        for y, this_creg in this_creg_dict.items():
+            # bundle
+            if this_creg['val'] > 1:
+                self.ax.plot([.6, .7], [y - .1, y + .1],
+                             color=self._style.cc,
+                             zorder=PORDER_LINE)
+                self.ax.text(0.5, y + .1, str(this_creg['val']), ha='left', va='bottom',
+                             fontsize=0.8 * self._style.fs,
+                             color=self._style.tc,
+                             zorder=PORDER_TEXT)
+            self.ax.text(-0.5, y, this_creg['label'], ha='right', va='center',
+                         fontsize=self._style.fs,
+                         color=self._style.tc,
+                         zorder=PORDER_TEXT)
+            self.ax.plot([0, self._cond['xmax']], [y, y], color=self._style.cc, zorder=PORDER_LINE)
+
+        # lf line
+        if feedline_r:
+            self._linefeed_mark((self._style.fold + 1 - 0.1,
+                                 - n_fold * (self._cond['n_lines'] + 1)))
+        if feedline_l:
+            self._linefeed_mark((0.1,
+                                 - n_fold * (self._cond['n_lines'] + 1)))
+
+    def _draw_ops(self, verbose=False):
+        _force_next = 'measure barrier'.split()
+        _wide_gate = 'u2 u3 cu2 cu3'.split()
+        _barriers = {'coord': [], 'group': []}
+        next_ops = self._ops.copy()
+        next_ops.pop(0)
+        this_anc = 0
+
+        #
+        # generate coordinate manager
+        #
+        q_anchors = {}
+        for key, qreg in self._qreg_dict.items():
+            q_anchors[key] = Anchor(reg_num=self._cond['n_lines'],
+                                    yind=qreg['y'],
+                                    fold=self._style.fold)
+        c_anchors = {}
+        for key, creg in self._creg_dict.items():
+            c_anchors[key] = Anchor(reg_num=self._cond['n_lines'],
+                                    yind=creg['y'],
+                                    fold=self._style.fold)
+        #
+        # draw gates
+        #
+        for i, (op, op_next) in enumerate(zip_longest(self._ops, next_ops)):
+            # wide gate
+            if op['name'] in _wide_gate:
+                _iswide = True
+                gw = 2
+            else:
+                _iswide = False
+                gw = 1
+            # get qreg index
+            if 'qubits' in op.keys():
+                q_idxs = op['qubits']
+            else:
+                q_idxs = []
+            # get creg index
+            if 'clbits' in op.keys():
+                c_idxs = op['clbits']
+            else:
+                c_idxs = []
+            # find empty space to place gate
+            if not _barriers['group']:
+                this_anc = max([q_anchors[ii].get_index() for ii in q_idxs])
+                while True:
+                    if op['name'] in _force_next or 'conditional' in op.keys() or \
+                            not self._style.compress:
+                        occupied = self._qreg_dict.keys()
+                    else:
+                        occupied = q_idxs
+                    q_list = [ii for ii in range(min(occupied), max(occupied) + 1)]
+                    locs = [q_anchors[jj].is_locatable(this_anc, gw) for jj in q_list]
+                    if all(locs):
+                        for ii in q_list:
+                            if op['name'] == 'barrier' and not self._style.barrier:
+                                q_anchors[ii].set_index(this_anc - 1, gw)
+                            else:
+                                q_anchors[ii].set_index(this_anc, gw)
+                        break
+                    else:
+                        this_anc += 1
+            # qreg coordinate
+            q_xy = [q_anchors[ii].plot_coord(this_anc, gw) for ii in q_idxs]
+            # creg corrdinate
+            c_xy = [c_anchors[ii].plot_coord(this_anc, gw) for ii in c_idxs]
+            # bottom and top point of qreg
+            qreg_b = min(q_xy, key=lambda xy: xy[1])
+            qreg_t = max(q_xy, key=lambda xy: xy[1])
+
+            if verbose:
+                print(i, op)
+
+            # rotation parameter
+            if 'params' in op.keys():
+                param = self.param_parse(op['params'], self._style.pimode)
+            else:
+                param = None
+            # conditional gate
+            if 'conditional' in op.keys():
+                c_xy = [c_anchors[ii].plot_coord(this_anc, gw) for ii in self._creg_dict]
+                if self._style.bundle:
+                    c_xy = list(set(c_xy))
+                    for xy in c_xy:
+                        self._conds(xy, istrue=True)
+                else:
+                    fmt = '{{:0{}b}}'.format(len(c_xy))
+                    vlist = list(fmt.format(int(op['conditional']['val'], 16)))[::-1]
+                    for xy, v in zip(c_xy, vlist):
+                        if v == '0':
+                            bv = False
+                        else:
+                            bv = True
+                        self._conds(xy, istrue=bv)
+                creg_b = sorted(c_xy, key=lambda xy: xy[1])[0]
+                self._subtext(creg_b, op['conditional']['val'])
+                self._line(qreg_t, creg_b)
+            #
+            # draw special gates
+            #
+            if op['name'] == 'measure':
+                vv = self._creg_dict[c_idxs[0]]['index']
+                self._measure(q_xy[0], c_xy[0], vv)
+            elif op['name'] == 'barrier':
+                q_group = self._qreg_dict[q_idxs[0]]['group']
+                if q_group not in _barriers['group']:
+                    _barriers['group'].append(q_group)
+                _barriers['coord'].append(q_xy[0])
+                if op_next and op_next['name'] == 'barrier':
+                    continue
+                else:
+                    if self._style.barrier:
+                        self._barrier(_barriers, this_anc)
+                    _barriers['group'].clear()
+                    _barriers['coord'].clear()
+            #
+            # draw single qubit gates
+            #
+            elif len(q_xy) == 1:
+                disp = op['name']
+                if param:
+                    self._gate(q_xy[0], wide=_iswide, text=disp, subtext='{}'.format(param))
+                else:
+                    self._gate(q_xy[0], wide=_iswide, text=disp)
+            #
+            # draw multi-qubit gates (n=2)
+            #
+            elif len(q_xy) == 2:
+                # cx
+                if op['name'] in ['cx']:
+                    self._ctrl_qubit(q_xy[0])
+                    self._tgt_qubit(q_xy[1])
+                # cz for latexmode
+                elif op['name'] == 'cz':
+                    if self._style.latexmode:
+                        self._ctrl_qubit(q_xy[0])
+                        self._ctrl_qubit(q_xy[1])
+                    else:
+                        disp = op['name'].replace('c', '')
+                        self._ctrl_qubit(q_xy[0])
+                        self._gate(q_xy[1], wide=_iswide, text=disp)
+                # control gate
+                elif op['name'] in ['cy', 'ch', 'cu3', 'crz']:
+                    disp = op['name'].replace('c', '')
+                    self._ctrl_qubit(q_xy[0])
+                    if param:
+                        self._gate(q_xy[1], wide=_iswide, text=disp, subtext='{}'.format(param))
+                    else:
+                        self._gate(q_xy[1], wide=_iswide, text=disp)
+                # cu1 for latexmode
+                elif op['name'] in ['cu1']:
+                    disp = op['name'].replace('c', '')
+                    self._ctrl_qubit(q_xy[0])
+                    if self._style.latexmode:
+                        self._ctrl_qubit(q_xy[1])
+                        self._subtext(qreg_b, param)
+                    else:
+                        self._gate(q_xy[1], wide=_iswide, text=disp, subtext='{}'.format(param))
+                # swap gate
+                elif op['name'] == 'swap':
+                    self._swap(q_xy[0])
+                    self._swap(q_xy[1])
+                # add qubit-qubit wiring
+                self._line(qreg_b, qreg_t)
+            #
+            # draw multi-qubit gates (n=3)
+            #
+            elif len(q_xy) == 3:
+                # cswap gate
+                if op['name'] == 'cswap':
+                    self._ctrl_qubit(q_xy[0])
+                    self._swap(q_xy[1])
+                    self._swap(q_xy[2])
+                # ccx gate
+                elif op['name'] == 'ccx':
+                    self._ctrl_qubit(q_xy[0])
+                    self._ctrl_qubit(q_xy[1])
+                    self._tgt_qubit(q_xy[2])
+                # add qubit-qubit wiring
+                self._line(qreg_b, qreg_t)
+            else:
+                logger.critical('Invalid gate %s', op)
+                raise QISKitError('invalid gate {}'.format(op))
+        #
+        # adjust window size and draw horizontal lines
+        #
+        max_anc = max([q_anchors[ii].get_index() for ii in self._qreg_dict])
+        n_fold = (max_anc - 1) // self._style.fold
+        # window size
+        if max_anc > self._style.fold > 0:
+            self._cond['xmax'] = self._style.fold + 1
+            self._cond['ymax'] = - (n_fold + 1) * (self._cond['n_lines'] + 1) + 1
+        else:
+            self._cond['xmax'] = max_anc + 1
+            self._cond['ymax'] = - self._cond['n_lines']
+        # add horizontal lines
+        for ii in range(n_fold + 1):
+            feedline_r = (n_fold > 0 and n_fold > ii)
+            feedline_l = (ii > 0)
+            self._draw_regs_sub(ii, feedline_l, feedline_r)
+        # draw gate number
+        if self._style.index:
+            for ii in range(max_anc):
+                if self._style.fold > 0:
+                    x_coord = ii % self._style.fold + 1
+                    y_coord = - (ii // self._style.fold) * (self._cond['n_lines'] + 1) + 0.7
+                else:
+                    x_coord = ii + 1
+                    y_coord = 0.7
+                self.ax.text(x_coord, y_coord, str(ii + 1), ha='center', va='center',
+                             fontsize=self._style.sfs,
+                             color=self._style.tc,
+                             zorder=PORDER_TEXT)
+
+    @staticmethod
+    def param_parse(v, pimode=False):
+        for i, e in enumerate(v):
+            if pimode:
+                v[i] = MatplotlibDrawer.format_pi(e)
+            else:
+                v[i] = MatplotlibDrawer.format_numeric(e)
+            if v[i].startswith('-'):
+                v[i] = '$-$' + v[i][1:]
+        param = ', '.join(v)
+        return param
+
+    @staticmethod
+    def format_pi(val):
+        fracvals = MatplotlibDrawer.fraction(val)
+        buf = ''
+        if fracvals:
+            nmr, dnm = fracvals.numerator, fracvals.denominator
+            if nmr == 1:
+                buf += '$\\pi$'
+            elif nmr == -1:
+                buf += '-$\\pi$'
+            else:
+                buf += '{}$\\pi$'.format(nmr)
+            if dnm > 1:
+                buf += '/{}'.format(dnm)
+            return buf
+        else:
+            coef = MatplotlibDrawer.format_numeric(val / np.pi)
+            if coef == '0':
+                return '0'
+            return '{}$\\pi$'.format(coef)
+
+    @staticmethod
+    def format_numeric(val, tol=1e-5):
+        abs_val = abs(val)
+        if isclose(abs_val, 0.0, abs_tol=1e-100):
+            return '0'
+        if isclose(fmod(abs_val, 1.0), 0.0, abs_tol=tol) and 0.5 < abs_val < 9999.5:
+            return str(int(val))
+        elif 0.1 <= abs_val < 100.0:
+            return '{:.2f}'.format(val)
+        return '{:.1e}'.format(val)
+
+    @staticmethod
+    def fraction(val, base=np.pi, n=100, tol=1e-5):
+        abs_val = abs(val)
+        for i in range(1, n):
+            for j in range(1, n):
+                if isclose(abs_val, i / j * base, rel_tol=tol):
+                    if val < 0:
+                        i *= -1
+                    return Fraction(i, j)
+        return None
